+++ src/components/manage/Form/Form.jsx
@@ -109,6 +109,7 @@ class Form extends Component {
     submitLabel: PropTypes.string,
     resetAfterSubmit: PropTypes.bool,
     isEditForm: PropTypes.bool,
+    isAdminForm: PropTypes.bool,
     title: PropTypes.string,
     error: PropTypes.shape({
       message: PropTypes.string,
@@ -132,6 +133,7 @@ class Form extends Component {
     submitLabel: null,
     resetAfterSubmit: false,
     isEditForm: false,
+    isAdminForm: false,
     title: null,
     description: null,
     error: null,
@@ -235,9 +237,10 @@ class Form extends Component {
 
   hideHandler = (data) => {
     return (
-      data['@type'] === 'text' &&
-      (!data.text ||
-        (data.text?.blocks?.length === 1 && data.text.blocks[0].text === ''))
+      !!data.fixed ||
+      (data['@type'] === 'text' &&
+        (!data.text ||
+          (data.text?.blocks?.length === 1 && data.text.blocks[0].text === '')))
     );
   };
 
@@ -779,6 +782,7 @@ class Form extends Component {
                               pathname={this.props.pathname}
                               block={block}
                               selected={this.state.selected === block}
+                              manage={this.props.isAdminForm}
                             />
                           </div>
                         </div>